refactor and img

+++ include/menu.h
@@ -2,8 +2,8 @@
 #define MENU_H
 
 void drawMenuBackground(GLuint textureID[]);
-void drawGameBackground(GLuint textureID[], int globalTranslationTotal);
-void drawBossBackground(GLuint textureID[], int globalTranslationTotal);
+void drawGameBackground(GLuint textureID[], float globalTranslationTotal);
+void drawBossBackground(GLuint textureID[], float globalTranslationTotal);
 void drawSelectionBox(GLuint textureID[], float posButton[], int selectedDifficulty);
 void drawButton(GLuint textureID[], float posButton[], int i, int id);
 void drawEndMenuButton(GLuint textureID[]);
